rename & relocate record test helper file
